+++ src/Routes/Challenge.ts
@@ -1,13 +1,11 @@
 import { Request, Response } from "express";
-import gm from 'gm';
+import sharp from 'sharp';
+import fetch from 'node-fetch';
 import { join, resolve } from 'path';
 import { ErrorCode } from "../Utility/Errors";
-import fetch from 'node-fetch';
 import { parse } from "url";
-import mime from "mime-types";
-import { getSize } from "../Utility/Image/size";
 
-const template = join(resolve('.'), 'src/templates'); // template directory
+const template = join(resolve('.'), 'src/templates/challenge.jpg'); // meme template
 
 export const Challenge = async (req: Request, res: Response) => {
     if(!('url' in req.query) || req.query.url.length === 0) {
@@ -23,11 +21,6 @@ export const Challenge = async (req: Request, res: Response) => {
             error: 'Bad URL',
             code: ErrorCode.BAD_URL
         });
-    } else if(!/(.png|.jpe?g)$/.test(url.href)) {
-        return res.status(400).send({
-            error: 'Bad URL',
-            code: ErrorCode.BAD_URL
-        });
     } else if(!url.host.endsWith('discordapp.net')) {
         return res.status(400).send({
             error: 'Bad URL',
@@ -35,74 +28,43 @@ export const Challenge = async (req: Request, res: Response) => {
         });
     }
 
-    const discord = await fetch(url.href);
-    const ext = mime.extension(mime.lookup(url.href) as string) as string;
-    const name = join(resolve('.'), `temp/challenge${Math.random().toString().slice(2)}.${ext}`);
-    
-    const buffer = await discord.buffer();
-    const size = await getSize(buffer);
-    const Image = gm(buffer); // downloaded image from Discord
-
-    if(size.width > 498 || size.height > 483 * .65) {
-        Image.resize(
-            size.width > 498 ? 400 : size.width,
-            size.height > 483 * .65 ? 375 : size.height
-        );
+    const resp = await fetch(url.href);
+    if(!resp.ok) {
+        return res.status(400).send({
+            error: `Received status ${resp.status} (${resp.statusText}).`,
+            code: ErrorCode.BAD_URL
+        });
     }
 
-    Image.write(name, err => {
-        if(err) {
-            return res.status(500).send({
-                error: err.toString(),
-                code: ErrorCode.SERVER_ERROR
-            });
-        }
-
-        const Comp = gm(join(template, 'challenge.jpg'))
-            .composite(name)
-
-        if(size.height >= 250) {
-            Comp.gravity('South')
-        } else {
-            Comp.gravity('Center');
-        }
-
-        Comp.toBuffer(ext, (err, img) => {
-            if(err) {
-                return res.status(500).send({
-                    error: err.toString(),
-                    code: ErrorCode.SERVER_ERROR
-                });
-            }
-
-            res.set('Content-Type', mime.lookup(url.href) as string);
-            return res.status(200).send(img);
+    const contentType = resp.headers.get('content-type');
+    if(!['image/png', 'image/jpeg'].includes(contentType)) {
+        return res.status(400).send({
+            error: 'Only PNG and JPG are accepted.',
+            code: ErrorCode.CONTENT_TYPE
         });
-    });
+    }
 
-    /*Image.write(name, err => {
-        if(err) {
-            return res.status(500).send({
-                error: err.toString(),
-                code: ErrorCode.SERVER_ERROR
-            });
-        }
+    const buffer = await resp.buffer();
+    const discord = sharp(buffer);
+    const size = await discord.metadata();
 
-        const Comp = gm(createReadStream(join(template, 'challenge.jpg')))
-            .composite(name)
-            .geometry('+xx+200')
-            //.gravity('center');
+    let resized: Buffer;
+    if(size.width > 498 || size.height > 483 * .65) {
+        resized = await discord.resize(
+            size.width > 498 ? 400 : size.width,
+            size.height > 483 * .65 ? 375 : size.height
+        ).toBuffer();
+    } else {
+        resized = buffer;
+    }
 
-        Comp.toBuffer(ext, (err, img) => {
-            if(err) {
-                return res.status(500).send({
-                    error: err.toString(),
-                    code: ErrorCode.SERVER_ERROR
-                });
-            }
+    const Image = await sharp(template)
+        .composite([{
+            input: resized,
+            gravity: size.height > 250 ? 'south' : 'center'
+        }])
+        .toBuffer();
 
-            res.set('Content-Type', mime.lookup(url.href) as string);
-            return res.status(200).send(img);
-        });
-    });*/
+    res.set('Content-Type', 'image/jpeg');
+    return res.status(200).send(Image);
 }