+++ src/Utility/Errors.ts
@@ -1,5 +1,6 @@
 export enum ErrorCode {
     MISSING_PARAMS = 'MISSING_PARAMS',
     SERVER_ERROR = 'SERVER_ERROR',
-    BAD_URL = 'BAD_URL'
+    BAD_URL = 'BAD_URL',
+    CONTENT_TYPE = 'CONTENT_TYPE'
 }