+++ src/Utility/Scheduled/Delete.ts
@@ -1,4 +1,0 @@
-/**
- * TO-DO: auto delete all images in ``temp`` directory
- * on some interval (5m, 10m, 60m, ??).
- */