+++ src/app.ts
@@ -1,9 +1,5 @@
 import express from 'express';
 import * as Routes from './Routes/';
-import { mkdirSync } from 'fs';
-import { resolve, join } from 'path';
-
-mkdirSync(join(resolve('.'), 'temp'), { recursive: true });
 
 const app = express();