+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # v1.0.0~beta1
 * Initial release.
 * Add in two templates (challenge & SpongeBob fire).
+
+# v1.0.0~beta2
+* Replace ``gm`` with ``sharp`` (libvips).
+* No longer needs to store images for ``challenge``.